refactor playlists functions
